+++ src/background/widgets/window_manager/state_v2.rs
@@ -25,7 +25,7 @@ use crate::{
     virtual_desktops::{events::VirtualDesktopEvent, SluWorkspacesManager2},
     widgets::window_manager::{
         cli::{Axis, NodeSiblingSide, StepWay},
-        handler::{schedule_window_position, set_app_windows_positions},
+        handler::{set_app_window_position, set_app_windows_positions},
         WindowManagerV2,
     },
     windows_api::{monitor::Monitor, window::Window},
@@ -268,7 +268,7 @@ impl TwmState {
         match side {
             TwmReservation::Float => {
                 self.add_to_floating(window, workspace_id);
-                set_rect_to_float_initial_size(window, &window.monitor()).log_error();
+                twm_set_rect_to_float_initial_size(window, &window.monitor()).log_error();
             }
             TwmReservation::Stack => {
                 let tree = self.get_or_insert_tree_mut(workspace_id);
@@ -322,7 +322,7 @@ impl TwmState {
         let residual = tree.add_to_tiled(window.address());
         for w in residual {
             tree.add_to_floating(w);
-            set_rect_to_float_initial_size(window, &window.monitor()).log_error();
+            twm_set_rect_to_float_initial_size(window, &window.monitor()).log_error();
         }
     }
 
@@ -333,7 +333,7 @@ impl TwmState {
                 let residual = tree.remove_window(&window_id);
                 for w in residual {
                     tree.add_to_floating(w);
-                    set_rect_to_float_initial_size(window, &window.monitor()).log_error();
+                    twm_set_rect_to_float_initial_size(window, &window.monitor()).log_error();
                 }
                 break;
             }
@@ -771,7 +771,7 @@ impl TwmState {
     }
 }
 
-pub fn set_rect_to_float_initial_size(window: &Window, monitor: &Monitor) -> Result<()> {
+pub fn twm_set_rect_to_float_initial_size(window: &Window, monitor: &Monitor) -> Result<()> {
     let guard = FULL_STATE.load();
     let config = &guard.settings.by_widget.wm.floating;
 
@@ -786,14 +786,13 @@ pub fn set_rect_to_float_initial_size(window: &Window, monitor: &Monitor) -> Res
     let x = monitor_rect.left + (monitor_width - window_width) / 2;
     let y = monitor_rect.top + (monitor_height - window_height) / 2;
 
-    schedule_window_position(
-        window.address(),
+    set_app_window_position(
+        window,
         Rect {
             left: x,
             top: y,
             right: x + window_width,
             bottom: y + window_height,
         },
-    );
-    Ok(())
+    )
 }