+++ changelog.md
@@ -14,6 +14,7 @@
 - fullscreen state being removed from background windows on foreground change.
 - dock and toolbar hidding while dragging items.
 - state while switching workspaces.
+- twm window not being moved when there is a fullscreen app on target monitor.
 
 ## [2.7.4]